# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Kino HTTP API
  summary: Stable, versioned contract for the Kino ROM library
  version: 0.2.0-preview.1
  description: |
    ROM catalog, metadata imports, managed media, and library maintenance.
    Protected operations require the configured owner bearer token.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
servers:
  - url: /api/v1
    description: Same-origin stable API entry point
security:
  - bearerAuth: []
...
paths:
...
  /series:
    get:
      operationId: getSeries
      tags: [Series]
      summary: List cross-platform series with member Games
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Projection'}
        - {name: q, in: query, schema: {type: string}, description: "Search series titles, descriptions, member titles, platforms, and relation types."}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated full or summary series, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/SeriesCollection'}, {$ref: '#/components/schemas/SeriesSummaryCollection'}]}}}}
    post:
      operationId: postSeries
      tags: [Series]
      summary: Create a cross-platform series and its optional complete member set atomically
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SeriesMutation'}
      responses:
        '201': {description: Created series, content: {application/json: {schema: {$ref: '#/components/schemas/Series'}}}}
        '400': {$ref: '#/components/responses/Error'}
...
  /games:
    get:
      operationId: getGames
      tags: [Library]
      summary: List games with editions and artifacts
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Projection'}
        - {name: q, in: query, schema: {type: string}}
        - {name: platform, in: query, schema: {type: string}}
      responses:
        '200': {description: Paginated full or summary games, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/GameCollection'}, {$ref: '#/components/schemas/GameSummaryCollection'}]}}}}
        '400': {$ref: '#/components/responses/Error'}
    post:
      operationId: postGames
      tags: [Library]
      summary: Create a game
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GameMutation'}}}}
      responses:
        '201': {description: Created game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/ImportConflict'}
...
  /sources:
    get:
      operationId: getSources
      tags: [Transfer]
      summary: List durable library source configurations
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated library sources, content: {application/json: {schema: {$ref: '#/components/schemas/LibrarySourceCollection'}}}}
    post:
      operationId: postSources
      tags: [Transfer]
      summary: Save a read-only source configuration below the library root
      requestBody: {$ref: '#/components/requestBodies/LibrarySourceRequest'}
      responses:
        '201': {description: Created source, content: {application/json: {schema: {$ref: '#/components/schemas/LibrarySource'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /source-scans:
    get:
      operationId: getSourceScans
      tags: [Transfer]
      summary: List source scan audit records
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: source_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated source scans, content: {application/json: {schema: {$ref: '#/components/schemas/SourceScanCollection'}}}}
...
  /hash-sources:
    get:
      operationId: getHashSources
      tags: [Transfer]
      summary: List imported ROM identity data sources and their active immutable releases
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated hash sources, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/HashSource'}}}}}}}
...
  /hash-packs/preview:
    post:
      operationId: postHashPacksPreview
      tags: [Transfer]
      summary: Validate and classify a privacy-minimized Kino hash pack without writing
      requestBody: {$ref: '#/components/requestBodies/HashPackUpload'}
      responses:
        '200': {description: Signed hash pack preview, content: {application/json: {schema: {$ref: '#/components/schemas/HashPackPreview'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
...
  /hash-packs/export:
    post:
      operationId: postHashPacksExport
      tags: [Transfer]
      summary: Export present ROM identities without ROM bytes, paths, saves, media, devices, or play history
      requestBody: {$ref: '#/components/requestBodies/HashPackExport'}
      responses:
        '200': {description: Kino hash pack ZIP, content: {application/vnd.kino.hashpack+zip: {schema: {type: string, format: binary}}}}
        '400': {$ref: '#/components/responses/Error'}
...
  /source-adapters:
    get:
      operationId: getSourceAdapters
      tags: [Transfer]
      summary: List versioned, declarative source adapter contracts
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated source adapters, content: {application/json: {schema: {$ref: '#/components/schemas/SourceAdapterCollection'}}}}
    post:
      operationId: postSourceAdapters
      tags: [Transfer]
      summary: Create a custom adapter that specializes an audited built-in import handler
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SourceAdapterMutation'}}}}
      responses:
        '201': {description: Created source adapter, content: {application/json: {schema: {$ref: '#/components/schemas/SourceAdapter'}}}}
        '400': {$ref: '#/components/responses/Error'}
...
  /frontend-adapters:
    get:
      operationId: getFrontendAdapters
      tags: [Runtime]
      summary: List versioned frontend adapter contracts
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated frontend adapters, content: {application/json: {schema: {$ref: '#/components/schemas/FrontendAdapterCollection'}}}}
    post:
      operationId: postFrontendAdapters
      tags: [Runtime]
      summary: Create a custom format bound to an audited compiled export handler
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FrontendAdapterMutation'}}}}
      responses:
        '201': {description: Created frontend adapter, content: {application/json: {schema: {$ref: '#/components/schemas/FrontendAdapter'}}}}
        '400': {$ref: '#/components/responses/Error'}
...
  /storage-cleanup/preview:
    post:
      operationId: postStorageCleanupPreview
      tags: [Storage]
      summary: Preview unreferenced files in service-managed ROM and media storage
      description: Scans only the service-managed roots. Referenced library or NAS paths, save blobs, active media staging files, symbolic links, and unsupported file types are never cleanup candidates. The signed token binds the complete candidate set.
      responses:
        '200': {description: Exact reviewable cleanup plan, content: {application/json: {schema: {$ref: '#/components/schemas/ManagedCleanupPreview'}}}}
        '409': {$ref: '#/components/responses/Error'}
...
  /media:
    get:
      operationId: getMedia
      tags: [Media]
      summary: List game- or edition-scoped media assets
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: game_id, in: query, schema: {type: string, format: uuid}}
        - {name: edition_id, in: query, schema: {type: string, format: uuid}}
        - {name: kind, in: query, schema: {type: string}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated media assets, content: {application/json: {schema: {$ref: '#/components/schemas/MediaAssetCollection'}}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Owner token or a server-issued device token; device scopes are enforced per route.
...
  parameters:
...
    Locale: {name: locale, in: query, schema: {type: string, enum: [zh-CN, zh-TW, ja, en]}, description: Preferred localized title. The resource default remains the fallback.}
    Projection: {name: projection, in: query, schema: {type: string, enum: [full, summary], default: full}, description: "Use summary for list metadata, aggregate artifact state, and an opaque cover reference; fetch the detail resource before editing or launching."}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}, description: Maximum number of records returned in this page.}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: Zero-based record offset in the endpoint's documented stable ordering.}
  requestBodies:
...
    LibrarySourceRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [name, kind]
            properties:
              name: {type: string}
              kind: {type: string, enum: [rom_directory, pegasus, esde, kino]}
              source_adapter_id: {type: string, description: Optional versioned adapter reference; defaults to the matching built-in handler}
              root_path: {type: string, description: "Relative ROM directory for rom_directory, or optional separate ROM content root for Pegasus and ES-DE; defaults to the metadata directory for metadata sources"}
              metadata_path: {type: string, description: "Relative metadata file for Pegasus, ES-DE, or neutral Kino v4/v5/v6 manifest"}
              runtime_metadata_path: {type: string, description: Optional explicit ES-DE es_systems.xml path; never auto-discovered or executed}
              platform: {type: string, description: "Required except for kino, whose entries carry independent platforms"}
              metadata_locale: {type: string, enum: [zh-CN, zh-TW, ja, en, '']}
              rom_storage_policy: {type: string, enum: [reference, copy], default: reference}
              media_storage_policy: {type: string, enum: [copy, reference, ignore], default: copy}
              enabled: {type: boolean, default: true}
...
    HashPackUpload:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [file]
            properties:
              file: {type: string, format: binary}
...
    HashPackExport:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [source_id, name, license, release]
            properties:
              source_id: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{0,127}$'}
              name: {type: string, maxLength: 200}
              publisher: {type: string, maxLength: 200}
              license: {type: string, maxLength: 200}
              release: {type: string, maxLength: 128}
...
  responses:
    ImportConflict:
      description: import_preview_stale when the signed snapshot no longer matches; import_batch_conflict when an atomic batch cannot commit; platform_definition_conflict/platform_definition_disabled/platform_key_conflict when a portable custom platform cannot be safely reused or created
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Kino-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Error:
      description: Structured API error
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Kino-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: invalid_argument
              message: request payload is invalid
              request_id: 40e39da4b75c41c28cf40b9f8de46c48
  schemas:
...
    HashSource:
      type: object
      required: [id, name, license, trust_level, record_count, release_count, created_at, updated_at]
      properties:
        id: {type: string}
        name: {type: string}
        publisher: {type: string}
        license: {type: string}
        trust_level: {type: string, enum: [local, imported, trusted]}
        active_release_id: {type: string, format: uuid}
        active_version: {type: string}
        active_pack_sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        record_count: {type: integer, minimum: 0}
        release_count: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    GameMutation:
      type: object
      additionalProperties: false
      required: [default_title, platform]
      properties:
        id: {type: string, description: Optional caller-generated identity on create; ignored on update.}
        default_title: {type: string, minLength: 1}
        platform: {type: string, minLength: 1, description: Canonical platform ID or registered alias.}
        titles: {type: object, additionalProperties: {type: string}}
...
    LibrarySource:
      type: object
      required: [id, name, kind, source_adapter_id, platform, rom_storage_policy, media_storage_policy, enabled, last_scan_status, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        kind: {type: string, enum: [rom_directory, pegasus, esde, kino]}
        source_adapter_id: {type: string}
        root_path: {type: string}
        metadata_path: {type: string}
        runtime_metadata_path: {type: string}
        platform: {type: string}
        metadata_locale: {type: string}
        rom_storage_policy: {type: string, enum: [reference, copy]}
        media_storage_policy: {type: string, enum: [copy, reference, ignore]}
        enabled: {type: boolean}
        last_scan_at: {type: string, format: date-time}
        last_scan_status: {type: string, enum: [never, scanning, ready, committed, failed, stale]}
        last_error: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    Game:
      type: object
      required: [id, platform, default_title, editions]
      properties:
        id: {type: string, format: uuid}
        platform: {type: string}
        default_title: {type: string}
        display_title: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        primary_edition_id: {type: string, format: uuid}
        editions: {type: array, items: {$ref: '#/components/schemas/Edition'}}
        media: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}
...
    SeriesMutation:
      type: object
      additionalProperties: false
      required: [default_title]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string, minLength: 1}
        description: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        members:
          type: array
          description: When present, replaces the complete member set in the same transaction as metadata.
          items: {$ref: '#/components/schemas/SeriesMemberMutation'}
    Series:
      type: object
      required: [id, default_title, display_title, titles, members, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        default_title: {type: string}
        display_title: {type: string}
        description: {type: string}
        titles: {type: object, additionalProperties: {type: string}}
        members: {type: array, items: {$ref: '#/components/schemas/SeriesMember'}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    GameCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Game'}}}}
    GameSummaryCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/GameSummary'}}}}
    SeriesCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Series'}}}}
    SeriesSummaryCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/SeriesSummary'}}}}
...
    LibrarySourceCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/LibrarySource'}}}}
    SourceScanCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/SourceScan'}}}}
    SourceAdapterCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/SourceAdapter'}}}}
...